Backend/JsonRpc: retry in 2 more cases of failed DNS resolution
Can't say for certain if it's my DNS provider or a Windows vs. Mac
difference, but I've had both of the problems addressed here cause
the test ``can connect to some electrum BTC servers`` to fail
consistently, the two problems addressed being 1) Dns.GetHostEntry
returning an empty AddressList for one of the IP Addressed in the
servers config file and 2) a SocketException with SocketError.NoData.
MSDN documentation describes this error as caused by:

"The usual example for this is a host name-to-address translation
attempt (using gethostbyname or WSAAsyncGetHostByName) which uses the
DNS (Domain Name Server). An MX record is returned but no A record”
indicating the host itself exists, but is not directly reachable."