Backend/JsonRpcSharp: new System.IO.Pipelines-based TCP client
Even though this one is still synchronous, it's a better foundation
for upcoming modifications (async, cancellation) than the old client
which had many quirks and was probably very slow. We don't shelve
the old client, though (we rename it to LegacyTcpClient), because
older versions of Mono have a runtime bug that is exposed when using
the new TCP client. It's not a big deal though, to have this version
used in the Ubuntu(Linux) client, because the upcoming improvements
for the new client are mainly targetted at mobile platforms, which
have more restrictions (and hence we need better cancellation story).

The Mono bug existing in v.4.6.2 manifests itself this way:
* Assertion at local-propagation.c:330, condition `ins->opcode > MONO_CEE_LAST' not met

Which looks to be the same as:
https://bugzilla.xamarin.com/show_bug.cgi?id=60625

Which seems to be fixed by:
https://github.com/mono/mono/pull/6020

Which was backported to 5.4:
https://github.com/mono/mono/pull/6242

( https://github.com/mono/mono/blob/d15-5-2017-06/configure.ac )