gitlab-ci,scripts: really build (not just restore) the GTK frontend
Follow-up to 4be1e13a0709998861a33dc7b1a66747373a16d3 to really build
the solution, not just restore the packages.

NOTE: There are three horrible workarounds included in this commit,
one to an already reported bug, and two that I haven't been bothered
to report yet (because it might be a bit tedious to come up with
self-contained testcases).