AzureDevOps Pipelines support
To make the solution build and run tests under new AzureDevOps
Pipelines (which is free with unlimited minutes for opensource
projects) we've had to do these 2 modifications:

a) Add one new way to find the Microsoft.FSharp.Targets file
(apparently the default VS2017 build agent host doesn't have
this installed, even if it has an F# compiler...). The best
workaround I've found for this problem[1] is getting it via nuget,
as suggested by @nosami.
Sadly, I've had to repeat this fix 3 times because xbuild (the
old mono tool equivalent to msbuild) doesn't have support for
Directory.Build.props[2] files (and we still need to support xbuild
if we want this to be compiled in Mono 4.6.2, the version included
by default in Ubuntu LTS), otherwise I would put all this probing
logic in one file...
(Also importing explicitly from a .props file didn't work either.)

b) Add a nuget dependency to NUnit's TestAdapter, so that the tests
can be understood by the VisualStudio tooling. See [3].

The motivation to switch from AppVeyor to AzureDevopsPipelines for
the Windows builds is that AppVeyor doesn't support Mac, while the
latter supports all platforms, and this way we will not need to have
3 different kind of CIs to support all platforms (now we don't need
to use Travis for macOS agents anymore).

[1] https://stackoverflow.com/questions/52418705/how-to-build-f-projects-with-azure-devops-pipelines-i-get-error-the-target-b
[2] https://docs.microsoft.com/en-us/visualstudio/msbuild/customize-your-build?view=vs-2017#directorybuildprops-and-directorybuildtargets
[3] https://stackoverflow.com/questions/40163527/how-to-run-nunit-test-in-visual-studio-team-services/40163778