Revert "Backend: upgrade NBitcoin to latest version (4.0.0.51 -> 4.1.1.48)"
This reverts commit 6cf138c99df7a4fb15556a3c15bb385030359000
(and subsequent commit 17c89f2ce3d9efd999b6228a7bf69a7bbc259d3d)
needed as a consequence of the first one (that broke the Windows
build).

The reason is that we're not ready yet to upgrade NBitcoin because
the PR https://github.com/MetacoSA/NBitcoin/pull/444 was not complete
(because it didn't open the API for .NET Standard, and we need that
for the frontend branch).