Backend: upgrade NBitcoin to latest version (4.0.0.51 -> 4.1.1.50)
This lets us drop the CryptSharpOfficial dependency
(thanks to https://github.com/MetacoSA/NBitcoin/pull/444),
but requires us to replace NBitcoin.Litecoin with the new
NBitcoin.Altcoins.

CryptSharpOfficial dep has some nuget problems when restoring
in the frontend branch, plus I trust it less than NBitcoin.

NOTE: the binding redirect is needed because otherwise we
would see exceptions similar to the one mentioned here[0].

(This was done already in [1], but we had to revert it [2]
because I found a problem with NBitcoin+.NETStandard that
I fixed in this other PR [3].)

[0] https://github.com/dotnet/corefx/issues/30642
[1] 6cf138c99df7a4fb15556a3c15bb385030359000
[2] c569bb2650a65bf9584d6b23b6452caff255de51
[3] https://github.com/MetacoSA/NBitcoin/pull/545