configure,Backend: switch the legacyTcpClient check to be at runtime
Due to the fact that we need a sufficiently-new Mono version to
build .NET Standard libraries (with msbuild), the only way to
build the GTK frontend is with new Mono version, even if the
result of the build would be employed later in a Linux box that
has the old Mono (4.6.2).

Therefore this check needs to be done at runtime instead of at
compile time.