Frontend.Console: bring the System.Net.Http crasher workaround
This crash (reported here [1]) was being workarounded in the
frontend branch, but it turns out the Console frontend (in
this master branch) is also affected, so we bring the workaround
here as well. This only happens in very old Mono versions
(older than 5.12), but the workaround can remain for all versions
because it doesn't seem to affect new versions of Mono.

[1] https://github.com/dotnet/corefx/issues/19914