Backend/Ether: ignore TLS auth problems if Mono not new enough
It turns out all ETC nodes we use/have need TLS1.2, which is not
available in earlier versions of Mono 4.8. The previous fix to
retry the connection in these cases just fixed the crasher, but
then the balance can never be retreived (it shows up as "?" in
the Console frontend). We try to ignore certificate errors then
to see if it helps.