Frontend.XF: if 1 currency balance can't be retreived, still show smth
Before this commit, if for any reason a currency (e.g. ETC in the case
of Linux+Mono4.6.2, as there's no TLS1.2 support there) balance couldn't
be retrieved, the total fiat balance shown would be ? USD, even if there
were some USD in other currencies.

Now we show the quantity sum of the balances of the currencies that
could be retreived, but we mark it with a "?" at the end, to let the
user know that "at least" he has XYZ amount of USD.