Backend/Ether: less no. of consistent responses required for ETC&Mono<4.8
Given that Mono 4.8 doesn't have TLS1.2 support and most of our
ETC servers require this, we're forced to be less consistent in
this case as a compromise. People that want better contrasted
server data (such as balances) should upgrade to newer versions
of Mono. (If we didn't do this change, Mono v<4.8 users would
get consistency-related crashes in FaultTolerantClient queries.)