Frontend.XF: add workarounds to make pages load in VS designer
VS iOS/Android designer needs empty constructors for the pages to
be able to show their layout. As we don't want these to be used
from our code but only from the designer, we add some weird
workarounds and hacks for this to happen (now that we have enabled
Warnings As Errors in the project).