src: make .fsproj files compatible again with VS4Windows
Strangely enough, when opening the solution in good old Visual
Studio 2017 for Windows (tested in versions 15.8.3 & 15.8.6)
the F# projects could not be opened.

To try to determine the reason, I created a new .fsproj from
scratch (for the Backend), and started adding items to it to
resemble the previous file. The change that made it break
again was the <Choose> structure that I had for detecting
where the FSharp.Targets file is.

Turns out, VS2017 creates .fsproj files that just use this
path[1], and it seems to work in Linux&Mac as well (tested in
VS4Mac 7.6.9.22, and CI in Linux+Mono is passing for both
Mono 4.6.2 and 5.x), don't ask me how. So we can simplify
all this now. And with this we can partially revert the
commit that added the FSharp.Compiler.Tools nuget package
as a workaround to make CI run on Azure Devops Pipelines[2].

[1] $(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\FSharp\Microsoft.FSharp.Targets
[2] da0e265961039bb64db7238984e37b95d7b5c545