ReadMe.md: change gitlab links from Diginex_HK to DiginexGlobal org
We had the 'Diginex_HK' name in Gitlab.com because 'Diginex' was taken,
and our twitter handle was 'Diginex_HK', but now our twitter handle has
switched to 'DiginexGlobal', so we're following this name change in
Gitlab.com as well.

I take this opportunity to extend the table to be long enough for the
longest row, and to specify the OS on each badge.