gwallet.sln: configure Frontend.XF.Android as deployable
This was needed to be able to run the Android frontend in the
emulator in Windows OS. Otherwise it would complain with the
error:

The project GWallet.Frontend.XF.Android needs to be deployed before it can be started. Verify the project is selected to be deployed in the Solution Configuration Manager.

(Change done in the menu Build->Configuration Manager.)