Backend: upgrade FSharp.Data to v3.0 final
Even though we didn't really need this version in the master branch,
(we needed it in the frontend branch because this version has
.NET Standard support, unlike the 2.x series), it's good to sync
this with the frontend branch because otherwise we keep getting
conflicts when merging dependency changes, especially now that
v3.0 final (non-beta, non-rc) has been released at last.

And checking if v3.0 just works is very easy: just run the console
frontend and see if the USD amounts display properly and there are
no crashes. (We mainly use FSharp.Data to parse the JSON queries we
get from the CoinMarketCap API to check for fiat prices.)