Frontend.XF: add workaround for macOS crash
As InsertPageBefore is not yet fully implemented in macOS,
it crashes. This is a wrong way of switching to a new page
(it introduces other problems, such as being able to go back
to the Loading page, which doesn't make any sense...) but
hey, at least it fixes the crash, which is the least worse
solution until the XamarinForms bug is fixed (and a new
version with it released).