Backend,Backend.Tests: use less decimals for crypto for bigger amounts
There was a UI problem with big crypto amounts, at least in the
Android frontend: the last word of the fiat balance (USD) and the
the crypto amount were not being shown, apparently because they
didn't fit and XamarinForms would crop the labels.

A workaround to this problem is use less decimals for crypto (by
default we had 5, which makes sense when the amount is very small)
if the amounts are bigger.