Frontend.XF: only create 10 Image instances
The previous code was creating new Image instances everytime the
user switched from hot to cold wallet view, which could aggrave
resource usage especially if there was a leak of some kind. With
this code we now instantiate 10 images (5 for hot, 5 for cold)
that will be used all the time.