Backend: fix build with older F#
Previous commit included a build fix for the new version of
VS for Windows 2017, which by the looks of it must be bringing
a new version of F# which broke the build. However, the fix
committed (in [1]) caused this compiler error in older F#:

error FS0256: A value must be mutable in order to mutate the contents or take the address of a value type, e.g. 'let mutable x = ...'

Using 'mutable' now we seem to achieve compatibility with both
versions of F#.

[1] 3be174e6405a35a45e91b99f0ddecb1d769893e0