Backend,Tests: allow for some slow servers to be prioritized up
Previous commit [1] gave an opportunity to non-history servers
or faulty ones to come earlier in the priority list (but not
earliest) in order to give them a chance to receive a good
score. This change goes in the same vein of that one, but:
- For all servers after the 4th position (and the 8th, and the
12th...)
- That is, not only the faulty and non-history ones.
This is to give an opportunity to slow servers to be remeasured
again and maybe raise in the list.

This commit is larger than I intended, because of two reasons:
- I needed to use a ListRemove feature, whose simpler way to
implement involved the need of having a way to identify
elements in the list (via equality). This caused a cascading
effect on the types that forced me to augment the Server<T,R>
type into Server<K,T,R> where K would be the TKey type that
controls equality.
- The above refactoring was going to be needed anyway, at some
point, because we will need to identify each server given to
the FaultTolerantParallelClient, so that we can save their
measured latency or fault in some cache file (this work will
be coming in upcoming commits).

[1] 1a1d53d926e45bf61b38d866b0dc977af3fd34a6