Backend,Tests: store less info for server faults
Storing the whole exception object was problematic because:
- It was making the cache file grow one order of magnitude in size
(and potentially making the caching code much less efficient).
- It was making the load of the cache file always throw some problem
that was causing the wallet to ditch it (probably the wallet thinking
it was receiving a DeserializationException due to a version mismatch,
but most likely because of some JSON parsing/encoding problem with
stacktraces characters).