Backend,Tests,Frontend.Console: introduce Mode.Fast and Mode.Analysis
Mode.Analysis is the normal mode of operation that tries to
maintain the server stat cache up to date (testing the ones
with no history first, and giving some small opportunity to
faulty and slower servers instead of leaving them behind all
the time with starvation). But Mode.Fast reaps the fruits of
the caching by having only the fast and non-faulty servers in
the beginning.

This fast mode will be used for all operations that are
required as a consequence of a manual intervention action
from the user (e.g. estimate fee, broadcast tx, etc.). And,
for frontends that have auto-refresh (all non console ones)
it will be used for the first balance check at startup (but
we still leave the balance refreshing queries for the Analysis
mode).