Backend: do less queries the very first time the wallet is run
When the wallet is run for the very first time (no caching file
existent) then there's not much point to check both confirmed and
unconfirmed balances. This way we're much faster checking balances
in this scenario.

To do this, I had to change the implementation of LoadFromDisk()
slightly, which now returns the firstRun bool flag and the map
(empty or not, instead of Option<Map<K,V>), which caused some
cascading changes in Caching.fs to make this compile again.