Backend,Frontend.XF: convert ImminentPayment from bool to Option<bool>
There were some cases in the code where we hardcoded a 'false'
value in cases where we were actually not sure if there's an
imminent payment or not (e.g. when we only check confirmed
balances). In these cases, it was better to return true so
that the balance refresh interval is shorter. However, if we
hardcoded 'true', code would become even more confusing, so
then I decided to convert this to an Option<bool>, in which
the NONE case means "we don't know".

When I refactored the code to change this and consequently
started to fix the cascading compilation errors that resulted
from this, I realised that I may have fixed a subtle bug that
was there in the previous code (when filter with .Any(), using
a match inside this lambda was not very readable and maybe it's
the reason the bug was there in the past, now I changed it to
use IsNone and IsSome, even if I recommended against it in the
coding style, because it's more readable).