Frontend.XF: if cached balance not available, get normal balance
To avoid adding more queries to normal app startup, we were querying
only cached balances for read-only accounts. However, it could happen
that:
- When pairing with the cold wallet, the balance queries were aborted
somehow.
- When updating to a new version of the wallet, cache file could be
destroyed because of version incompatibility of the file.
This was causing a crash in these cases:
https://sentry.io/nblockchain/gwallet/issues/817417238/

Now, as a last resort, we query balances normally (slower, but at
least it will not crash).