Backend: first pass at using TransactionBuilder for btc/ltc transfers
I was delaying this piece of work as much as possible because
it was a tricky migration (and actually, this is not finished,
that's why I call this just a "first pass"), however recently
I received an exception when trying to send BTC that told me
that the dirty fee estimation algorithm hadn't succeeded at it.

This was an exception instead of a warning, so as to force me
to fix it. So I did, but this required a big refactoring,
mainly because estimating fees accurately with the
TransactionBuilder needs the redeemScript (see [1] for more
info), which can be obtained with the public key, so I had
to overhaul the cold storage mode to obtain the public key,
instead of the public address, when pairing.

This involved changing how ReadOnlyAccounts are stored on
disk. And it also involved changing the JSON format of the
BTC/LTC transactions, that's why I'm bumping the version
from 0.3.0.0 to 0.3.1.0.

But overall I'm very happy with this change because it
removed some FIXMEs, the huge dirty fee-estimation hack,
and last but not least, I improve the serialization and
deserialization tests a bit more to avoid hardcoding JSON
in F#, but to use yet another embedded .json file.

[1] https://github.com/MetacoSA/NBitcoin/issues/624