Frontend.XF.Android: retarget API versions: min->15, target->26
This was what I should have done in this commit [1], but
this bug[2] prevented me to do it properly. Now we have
a sane minimum Android version so that more people can
install the app.

Credit goes to Andrei <andrei.misiukevich@gmail.com> for
trying to install the app on his Android and telling me
about this attribute when he couldn't install it.

[1] ed6d60291ae93c3985b1616eb89588cdc59b9734
[2] https://github.com/mono/monodevelop/issues/6846