Backend/UtxoCoin: add dirty workaround for TxBuilder+AllBalance bug
Recent commit refactoring to use TransactionBuilder[1] introduced a
new bug[2] when all balance needs to be sent. A proper fix is in the
works, but in the meantime this can be workarounded by trying to call
many times to EstimateFees(), catching the exception and keep increasing
the fee until there is no exception.

[1] c96dd339bed1c573b0db707db8dc51896ba48894
[2] https://gitlab.com/DiginexGlobal/geewallet/issues/45