Revert "Frontend.XF: stop using Task.ContinueWith(), in almost everywhere"
This reverts commit 07e79119c087dcc2cf26090469cac8d2ae358d37 because
it causes unexpected hangs (e.g. after scanning a QR code).

By the looks of it, you may think that the culprit of the above problem
might be that this patch introduced nested BeginInvokeOnMainThread calls
but:
- Even after removing the clearest instances of this (the methods
AssignScannedQrCodeProperties and AssignScannedQrCodeProperty), the
hang still happens.
- I tried to create a helper method FrontendHelpers.InvokeOnMainThread
that checks Device.IsInvokeRequired first (if this returns false, AFAIU
it means we're already in the main thread, and then the action can be
executed directly), but it didn't help.

So this requires more investigation.