Backend/JsonRpcSharp: fix nonLegacy (async&sysIOpipelines) TCP client
Fixes https://gitlab.com/DiginexGlobal/geewallet/issues/54

Config: enable new non-Legacy TCP client again

It works in Windows, Android, iOS, and Linux(if Mono>=5.18).

In macOS it doesn't work (still returns empty string, same
bug that Evgeny fixed for the majority of the platforms).

In very old Mono (4.6.2 or similar), still throws an unmanaged
runtime crash.
In old Mono (older than 5.18 but not as old as 4.6.2) we have
the same issue as in macOS: empty string received.

So let's use the LegacyTcpClient for all the cases where the
new one doesn't work.

NOTE: The IsMacPlatform() function is a bit convoluted because
it turns out that on iOS platform, Environment.OSVersion was
also reporting Platform.Unix and it all the mac folders were
also existing. So let's use the new Xamarin.Essentials nuget
to check if we're on iOS (if we're not, then we assume we're
on Mac).

Closes https://github.com/diginex/geewallet/issues/54

Pull request: https://github.com/diginex/geewallet/pull/55