Backend/Config,JsonRpcSharp: enable new TCP client again
It works in Windows, Android, iOS, and Linux(if Mono>=5.18).

In macOS it doesn't work (still returns empty string, same
bug that Evgeny fixed for the majority of the platforms).

In very old Mono (4.6.2 or similar), still throws an unmanaged
runtime crash.
In old Mono (older than 5.18 but not as old as 4.6.2) we have
the same issue as in macOS: empty string received.

So let's use the LegacyTcpClient for all the cases where the
new one doesn't work.