Backend,Frontend.Console: workaround ETH token estimation issue
After filing a bug about wrong DAI fees estimation[1], I was adviced
in the Nethereum channel that this could be because of a buggy node.
As we don't control the nodes, it's better to just not store the
estimated fee as outgoing transaction, and wait for the balance
to update from an unconfirmed-balance query (it's just the miner
fee anyway, we're not in such a hurry of updating the balance in
this case, especially because the currency being transferred is
not ETH, so the user might not even be looking at it).

If we don't do this, when the new balance comes, the outgoing
transaction stored in cache for the fee would not be deleted,
because the amount would not match, and therefore the amount
would still be substracted when showing it to the user, even
if we get a proper balance coming from the network (thus showing
an incorrect balance instead of a not-up-to-date one, which is
not desirable in any case).

[1] https://github.com/Nethereum/Nethereum/issues/509