bump(downgrade) to v.0.2.1.1
I just did the following:
- cloned 'stable' into 'oldstable' (which had a 0.1 version)
- deleted 'stable'
- cloned 'master' (which has a 0.3 version) into 'stable'
(to promote a new stable checkpoint).
But in order to differentiate the new 'stable' from 'master' I
either need to upgrade master, or downgrade stable. I choose
the latter because it's less work (as we have labeled the
versions this way in the ReadMe.md).