Backend: speedup balance querying taking advantage of caching
This approach to speed up the retreival of balances is a fast-path
approach that takes in account:
- The cached value of the balance for the same account, regardless
of when it was taken.
- The unconfirmed balance first (in Ether-like currencies, this is
faster to retreive than the confirmed balance, as the latter requires
two queries).

If both cached balance and unconfirmed balance can be retrieved (fast
path), and the latter is lower or equal to the cached balance, we can
already show this up to date value.