Backend: speedup UxtoCoin balance retreival by doing less queries
Electrum servers (in the UtxoCoin world) return both confirmed and
unconfirmed balances at the same time in the same query; but in the
Ether world, we need to do 2 queries to retrieve the confirmed
balance of an account, and 1 query to retreive the unconfirmed one.

The previous commit[1] took advantage of caching for a fast-path
in which we would only need to query unconfirmed balances, but this
is only advantageous in the ether world, as explained in the previous
paragraph. So then I move this optimized logic to the EtherAccount
module and I leave the UtxoCoin with a simpler way to retreive the
balance.

This change is also a speedup for the UtxoCoin-based balances (BTC,
LTC), because now we don't do 2 queries for retreiving the confirmed
and unconfirmed balances, but just 1.

[1] 32ffc72fa551a04e72fc2c57ecb564fca3adb1cb