Backend,Frontend.Console: gather server stats from github or gitlab
Now that we committed an example file[1] of server stats gathered by
me running the wallet, we can download it from github or gitlab when
the user is about to create the accounts (the private key). This
way we can be much faster querying balances when running the wallet
for the first time.

[1] 2b1052b1259b62a930d3d957edecc878301cd05f