Frontend.Console: crash if background bootstrap task crashes
As Console applications don't really have a way to call back into
the main thread[1] (something which would allow me to create some
helper method, like frontend's branch FrontendHelpers.DoubleCheck*
methods, which make the whole app crash if a child task crashes),
then the best way to check that the task finished properly is to
manually verify that its .Exception property is null or not.

(If this makes the app crash, it would be picked up by Sentry, in
theory, so we know about this. Much better than having a task fail
silently which BTW seems unavoidable in Mono, as it seems that the
TaskScheduler.UnobservedTaskException[2] delegate or the
ThrowUnobservedTaskExceptions[3] config settings are not working
at all even if Alex seems to have implemented them[4][5] some
time ago. TODO: report this.)

[1] https://stackoverflow.com/a/31318005/544947
[2] https://docs.microsoft.com/en-us/dotnet/api/system.threading.tasks.taskscheduler.unobservedtaskexception
[3] https://docs.microsoft.com/en-us/dotnet/framework/configure-apps/file-schema/runtime/throwunobservedtaskexceptions-element
[4] https://bugzilla.xamarin.com/show_bug.cgi?id=36560
[5] https://github.com/mono/mono/pull/2424