Revert "Backend,Tests,Frontend.Console: 4.6.1 & use JsonRpcSharp"
This reverts commit d7cf73bc6baa4838a5a922b53b08981208be0731
and commit 5edf5517444aa1dea1d5a7cb83ec32d255a5f301.

Because it breaks the Windows build (in particular the tests
run), maybe due to using .NET4.6 nuget dependencies after we
have upgraded the projects to 4.6.1 (Windows' assembly loader
has always seemed to be more restrictive to me, and here it
may find that JsonRpcSharp's nuget links to a 4.6.1 assembly
which is different than the one that GWallet.Backend links to).