Backend/Config: increase number of retries for fault tolerance
One unintended consequence of the recent drastic timeout reduction[1]
is that we get ResultInconsistencyException[2]s more often.

Let's alleviate this problem a bit by increasing the number of
retries to same servers, until we find a better solution (which
would probably involve having the FaultToleranceClient control
the timeout, starting low, and increasing it everytime it gets
a NoneAvailable or ResultInconsistencyException, until a certain
MAX_RETRIES is reached).

[1] 2adad42506699a7ea4ae0fca539f3987a7e46553
[2] https://sentry.io/organizations/nblockchain/issues/884791785/