Backend/Config: increase network timeout from 3s to 30s
Overall, the drastic timeout reduction[1] caused two problems:
* More frequent ResultInconsistencyExceptions, a problem which
was mitigated in [2].
* Some currencies' (especially ETH since their servers seem
overloaded) balances sometimes could not be retreived (and showed
as ? or * in the UI).

For the latter problem we will need to have two timeout settings:
* One, higher, that will kill the thread if it expires, deeming
it as too slow server.
* Another one, much lower, which will just make FaultTolerantClient
spawn more tasks to try to retreive the balance of the currency
that could not be retreived after this time.

But until we have the above fix, we mitigate yet again this
problem by increasing this timeout again.

[1] 2adad42506699a7ea4ae0fca539f3987a7e46553
[2] ff1cde167d55602b6215ba523315f618d15c0b40