Backend(,Tests): cancel all FaultTolerantParallelClient unfinished jobs
This doesn't mean that we've already hooked up into the CancellationToken
APIs of JsonRpcSharp and Nethereum (that will come in the next commit),
but at least by dropping the Ignore() attribute of this test, we know
that the token is effectively canceled. We now just need to propagate it
(mainly to the C# end, cause in F# it should be propagated automatically).