scripts/bump.fsx: use odd minorRevisions for unstable, even for stable
From now on, we will only tags versions whose minorRevision is even
(e.g. 0.4.1.20, 0.3.2.88), and leave odd versioms for unreleased and
untagged versions (e.g. 0.4.1.21, 0.3.2.89).

This is important because we package some artifacts with the version
number in the artifact name, and it will allow us to know if the
artifact (e.g. a zip file) comes from a released version (even number)
or from an odd version (which means it's from some commit that was
pushed in between the previous and the next even version). So if you
see an artifact such as 0.3.3.3, it means you cannot determine exactly
to which commit hash it exactly belongs, but if it's 0.3.3.4, you know
that this version number maps to a git tag in the repository (and thus
a commit hash).