Backend,Frontend.Console: move out of gas checks to the backend
It's better implemented this way, for two reasons:
- By throwing an exception, we don't let the call to
SaveOutgoingTransactionInCache call to happen, therefore no
balance is substracted if transaction runs out of gas (although
in this case we should substract the fee, so I added a FIXME
about this).
- Less code duplication across frontends: we will only need a
try-with block in each frontend to find the possible issue with
the transaction.