Backend(,.Tests),Frontend.Console: finish cancellation support
Recent commits introduced cancellation support when using the
FaultTolerantParallelClient (either to cancel the unfinished
jobs itself or to allow cancelling them from an external source).

This finishes the job by:
- Upgrading Nethereum version to use the latest cancellationToken
API additions.
- Propagating the cancellationToken to all Web3 queries (this is
needed because the cancellationToken only gets propagated in an
automatic way when staying with F#'s Async workloads, but by using
Nethereum we're crossing the boundary from F# to C#).
- Allowing to provide a cancellationTokenSource object on the
Backend.Account queries related to checking the balances, because
they are the only ones (so far) that get run in the background in
the XF frontend, so they need to be killed when switching out from
the app (the equivalent of ALT+TAB in desktop) because mobile OSs
are very restrictive about resource usage when the app is not in
the foreground.
- Adding exploratory unit tests that I created to understand better
how cancellation works with Async<'T> and Task<'T> elements.