Backend/FaultTPClient: try to fix race condition introduced previously
In the previous commit we started printing whole stacktraces for
exceptions that happened when the cancellation had already been
requested. However, it could happen that this exception is thrown
roughly around the same time that the cancellation is requested.
Therefore we now use a little capsule of mutable state that lets us
know if the cancellation has been requested recently or not. If it
has not been called recently (e.g. already cancelled 4seconds ago)
we definitely want to report the issue.