Backend,Frontend.Console: don't allow sending to contract addresses
The wallet doesn't support this yet (best proof of this is that when
you send money with it, the transaction runs out of gas).

Even if we implemented "out of gas detection" in [1], it's better to
prevent the transaction to be sent in the first, place so that the
user doesn't spend the gas (ether value) required to find out that
the transaction will ran out of gas.

Fixes https://gitlab.com/DiginexGlobal/geewallet/issues/69

[1] 583cee592c34018f0c8c9df9e588bb427cd6ef44