Backend: prevent bypassing Ether address full-validation
We recently implemented contract-address detection at validation
time[1] but I just found a case when this prevention wouldn't
work: in case of using a contract address whose checksum was
incorrect, because a previous validation check would kick in
before this one. With this commit, we check this also in this
case.

[1] 245e06c568b16c457a8ffe4d178de7f1e0f278e3