Backend: adjust marshalling code to fail faster
See the URLs in the FIXME we had for more info. This was easier
than I thought (at first, I thought it was breaking all tests,
but that was just because I switched Stratum impl to use these
default settings, which is wrong because JsonRpc assumes a style
of members based on lower_case notation, not PascalCase).