Backend(,Tests): massive win for the FaultTPClient
I found a corner case in which the max number of parallel jobs was not
being honored so I wrote a test for it. Then, when trying to fix it, I
found a massively smart way to refactor the code that makes it so much
simpler than before, and it fixes the corner case too! I <3 myself lol