Frontend.XF: create frame in LoadingPage
This is necessary for our next step which will be just making the
frame visible when the currency balance is retreived (showing the
balances page as soon as the first balance is retreived).

For this refactoring we've had to make a small sacrifice: currency
image preloading is no longer parallel, however this should be a
very short time and is still done while the "Loading..." label is
being shown.