Backend: implement yet another speed improvement (for first run)
We implemented recently a speed improvement that consisted in only
querying one server (consistency of data configured as 1) in the
case that the first value retrieved from the first server matches
with what we have in cache. The problem is that the first time the
app is run, there are no cached values, so this wouldn't help in
this scenario. However, we can consider that the likelyhood of
having balance higher than zero in the first run of the app very
unlikely (still possible if someone is simply recovering their
funds from their seed that contained previous funds, but for this
unlikely event we consider the chance of being cheated by a server
being low).

This way, we optimize the scenario when running the wallet for the
first time \o/