Backend: simplify and optimize versioning and version reporting
Turns out I could rename "gwallet" to "geewallet" in Sentry, but anyway
the project name doesn't seem to be needed even if it's documented to be
sent this way. Last but not least, by making this a constant instead of
a function, we make it's access much faster for the subsequent times that
it'll be called (sure this might be only a micro-optimization...).