Backend: decrease number of parallel jobs
After implementing the progress bar that hints to us how fast is each
currency, we can see that in most cases the slowness is not really
caused by a very slow currency compared to the rest, but somehow most
currencies take the same amount of time to load, and the slow part seems
to be for the jobs to start (e.g. about 5-10 seconds with no balances of
any currency loaded at all, and suddenly in less than 1 second all
currencies load).

This makes me think that maybe we're giving too many jobs to the app
to chew at the same time, so let's try decreasing this number a lot
to see if it improves things.