Backend: change branch to bootstrap server data from
I'm going to make this part of the release process (one of the things
that bump.fsx does) to make this easier and not forget, for every
release. So, to make this more manageable without having to merge
branches, I'm just going to run 'update-servers' makefile target in
the 'frontend' branch.