Backend: overhaul FaultTolerantParallelClient's cancellation
Hopefully the last attempt to fix the weird cancellation bug [1][2]
(that currently doesn't crash the app anymore, but generates a warning
and marks the balance with '?' to mean 'unknown').

This change makes it so that there's independent cancellationTokenSources
for each job, instead of having one shared across the FaultTPClient.Query
work, because I suspected that one job running HttpClient and facing the
TCE (to denote a timeout) could make the other jobs be cancelled as well.

[1] https://gitlab.com/knocte/geewallet/issues/125
[2] https://github.com/dotnet/corefx/issues/20296 ?