Backend/FaultTolerantParallelClient: possible final fix for TCE issue
I think I just found the culprit for the deadly TCE problem I've been
having for the past few weeks/months:
https://gitlab.com/knocte/geewallet/issues/125

The root cause might have been this .NETCore's BCL bug: [1].

However, after working on a workaround (which I committed here[2]),
I was still getting cancellation[3] in code that wasn't supposed to
cancel at all (at least from an external cancellationSource, since
this only is supposed to happen when the user sends the app to the
background while it is in the middle of refreshing balances; unlike
what happens with the first retreival of balances on the Loading page).

After committing the workaround, I started ironing out issues but which
were just derived from the new workaround approach, while still overlooking
the real root cause, which I just realised is the following: balance
queries in the Ether world (i.e. ETH, ETC, DAI) need sometimes 2 distinct
queries (not just one), to retreived both the confirmed and the unconfirmed
balance.

This was the reason why I was only reproducing this problem in this kind
of currencies (which led me to believe the real culprit was the HttpClient
bug[1], because the UTXO-based ones don't use HttpClient...). And another
reason why this was hard to reproduce is because wallets that already have
cached balances used them as a reference point (against unconfirmed balance
amount) instead of querying for both unconfirmed&confirmed balances, so
most people would reproduce this only the first time they used geewallet.

The fix (see the diff of this commit) might look too simple, but it would
have not been possible without the change in [2] (and especially, [4]). It
basically consists of avoiding to dispose(or cancel) the cancelSource from
within FaultTolerantParallelClient itself, and only let external code do
it. This way, finishing a confirmed-balance-query would not cancel an
unconfirmed-balance-query for the same account, and viceversa.

I was actually going to commit this change just to dismiss this issue
altogether and hide the shit under the rug, but the only thing that needed
to be done besides the change is *UNDERSTAND* why it's needed, hence this
long commit message.

[1] https://github.com/dotnet/corefx/issues/20296
[2] 54b2a367a114856d5bdcbfef344c94cf39c58d78
[3] e.g. https://sentry.io/organizations/nblockchain/issues/1353874620
[4] ab586a66c9f6f7e53aae84358bd4cd83c74ae99b