Backend: Tolerate low amount of servers
What: Tolerate less than the value of NumberOfParallelJobsAllowed
amount of servers for a coin We use the funciton "min" to ensure
we never take more than the available amount of servers from a list.

Why: If there are not many servers configured, we will fail
generating requests for the few servers available.
NumberOfParallelServers is currently set to 5, and the following
line (in the changed source file) throws an exception if you ask
for more elements than the list contains (e.g. asking for 5 when
the list contains 4 elements will throw).