Frontend.XF(CircleChartView): enable AntiAlias for the pie
I just saw this paragraph in theee docs:
"The SKPaint class defines an IsAntialias to enable anti-aliasing in
the rendering of your graphics. Anti-aliasing generally results in
visually smoother edges, so you'll probably want to set this property
to true in most of your SKPaint objects. For purposes of simplicity,
this property is not set in most of the sample pages."

From https://docs.microsoft.com/en-us/xamarin/xamarin-forms/user-interface/graphics/skiasharp/basics/circle